Created new DTO for request/response. Removed unnecessary CardController. Role entity replace by RoleEnum. Constants moved to application.properties. Attempt to separate logic in security filters had made. And other small improvements had added.

+++ src/main/java/ru/dubovitsky/flashcardsspring/repository/CardSetRepository.java
@@ -3,5 +3,6 @@ package ru.dubovitsky.flashcardsspring.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.dubovitsky.flashcardsspring.model.CardSet;
 
-public interface CardSetRepository extends JpaRepository<CardSet, Integer> {
+public interface CardSetRepository extends JpaRepository<CardSet, Long> {
+
 }
